First attempt at basic e2e
This one auths with Spotify, finds the user's most recent albums, and
adds the tracks from the most recent album to the queue. It's close to
being able to play one when clicked.

But, the architecture is a mess, and trying to get the instance of the
SpotifyAppRemote to the click handler is way too hard. After doing some
reading about how to architect an Android app, it's time to scrap the
whole thing and try again.